requirements.txt
gspread~=5.0.0
PyYAML~=5.3.1
requests~=2.24.0
python-telegram-bot